Move Account component into its own file
